[#] - Refactor de la carpeta tests
